A  bunch of fixes for proper and smooth operation under Windows and .NET 2003
git-svn-id: https://xlr.svn.sourceforge.net/svnroot/xlr/trunk@29 ed21b5e9-bf16-0410-b9ea-f1dd0686f206

+++ xl2/errors.cpp
@@ -58,7 +58,7 @@ void XLError(XLErrorNumber err, text file, uint line,
     {
         char buffer[10];
         sprintf(buffer, "$%d", i+1);
-        uint found = uint(errMsg.find(buffer));
+        size_t found = errMsg.find(buffer);
         if (found)
             errMsg.replace(found, strlen(buffer), args[i]);
     }
